+++ ui/src/routes.tsx
@@ -7,7 +7,7 @@ import { useUser } from './context'
 import { LandingPage } from './pages/LandingPage'
 import { Home } from './pages/Home'
 import { BasePage } from './pages/BasePage'
-import { UserProfilePage } from './pages/UserProfilePage'
+import { UserProfilePage } from './pages/user_profile_page/UserProfilePage'
 import { ForgotPasswordPage } from './pages/ForgotPasswordPage'
 import { ResetPasswordPage } from './pages/ResetPasswordPage'
 import { VerifyEmailPage } from './pages/VerifyEmailPage'
